reorganizing widget files
+ fixed frame size for widgets
